pyproject.toml
[project]
name = "audiostat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "flake8>=7.3.0",
    "mypy>=1.20.0",
    "winrt-windows-foundation>=3.2.1",
    "winrt-windows-media-control>=3.2.1",
]
